feat: implement remote signing transport for S3 requests with chunked encoding handling
